@@ -4,7 +4,7 @@ version: 2.0.0
 homepage: https://github.com/rmtmckenzie/flutter_qr_mobile_vision
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0 <2.0.0"
 
 flutter:
@@ -19,4 +19,4 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  native_device_orientation: "^0.4.3"
+  native_device_orientation: ^1.0.0
