@@ -6,53 +6,55 @@ import 'package:flutter/rendering.dart';
 import 'package:native_device_orientation/native_device_orientation.dart';
 import 'package:qr_mobile_vision/qr_mobile_vision.dart';
 
-final WidgetBuilder _defaultNotStartedBuilder = (context) => new Text("Camera Loading ...");
-final WidgetBuilder _defaultOffscreenBuilder = (context) => new Text("Camera Paused.");
-final ErrorCallback _defaultOnError = (BuildContext context, Object error) {
+final WidgetBuilder _defaultNotStartedBuilder =
+    (context) => Text("Camera Loading ...");
+final WidgetBuilder _defaultOffscreenBuilder =
+    (context) => Text("Camera Paused.");
+final ErrorCallback _defaultOnError = (BuildContext context, Object? error) {
   print("Error reading from camera: $error");
-  return new Text("Error reading from camera...");
+  return Text("Error reading from camera...");
 };
 
-typedef Widget ErrorCallback(BuildContext context, Object error);
+typedef Widget ErrorCallback(BuildContext context, Object? error);
 
 class QrCamera extends StatefulWidget {
   QrCamera({
-    Key key,
-    @required this.qrCodeCallback,
+    Key? key,
+    required this.qrCodeCallback,
     this.child,
     this.fit = BoxFit.cover,
-    WidgetBuilder notStartedBuilder,
-    WidgetBuilder offscreenBuilder,
-    ErrorCallback onError,
+    WidgetBuilder? notStartedBuilder,
+    WidgetBuilder? offscreenBuilder,
+    ErrorCallback? onError,
     this.formats,
   })  : notStartedBuilder = notStartedBuilder ?? _defaultNotStartedBuilder,
-        offscreenBuilder = offscreenBuilder ?? notStartedBuilder ?? _defaultOffscreenBuilder,
+        offscreenBuilder =
+            offscreenBuilder ?? notStartedBuilder ?? _defaultOffscreenBuilder,
         onError = onError ?? _defaultOnError,
-        assert(fit != null),
         super(key: key);
 
   final BoxFit fit;
-  final ValueChanged<String> qrCodeCallback;
-  final Widget child;
+  final ValueChanged<String?> qrCodeCallback;
+  final Widget? child;
   final WidgetBuilder notStartedBuilder;
   final WidgetBuilder offscreenBuilder;
   final ErrorCallback onError;
-  final List<BarcodeFormats> formats;
+  final List<BarcodeFormats>? formats;
 
   @override
-  QrCameraState createState() => new QrCameraState();
+  QrCameraState createState() => QrCameraState();
 }
 
 class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
   }
 
   @override
   dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     super.dispose();
   }
 
@@ -72,7 +74,7 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
   }
 
   bool onScreen = true;
-  Future<PreviewDetails> _asyncInitOnce;
+  Future<PreviewDetails>? _asyncInitOnce;
 
   Future<PreviewDetails> _asyncInit(num width, num height) async {
     var previewDetails = await QrMobileVision.start(
@@ -111,14 +113,16 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
 
   @override
   Widget build(BuildContext context) {
-    return new LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
+    return LayoutBuilder(
+        builder: (BuildContext context, BoxConstraints constraints) {
       if (_asyncInitOnce == null && onScreen) {
-        _asyncInitOnce = _asyncInit(constraints.maxWidth, constraints.maxHeight);
+        _asyncInitOnce =
+            _asyncInit(constraints.maxWidth, constraints.maxHeight);
       } else if (!onScreen) {
         return widget.offscreenBuilder(context);
       }
 
-      return new FutureBuilder(
+      return FutureBuilder(
         future: _asyncInitOnce,
         builder: (BuildContext context, AsyncSnapshot<PreviewDetails> details) {
           switch (details.connectionState) {
@@ -130,11 +134,11 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
                 debugPrint(details.error.toString());
                 return widget.onError(context, details.error);
               }
-              Widget preview = new SizedBox(
+              Widget preview = SizedBox(
                 width: constraints.maxWidth,
                 height: constraints.maxHeight,
                 child: Preview(
-                  previewDetails: details.data,
+                  previewDetails: details.data!,
                   targetWidth: constraints.maxWidth,
                   targetHeight: constraints.maxHeight,
                   fit: widget.fit,
@@ -142,17 +146,17 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
               );
 
               if (widget.child != null) {
-                return new Stack(
+                return Stack(
                   children: [
                     preview,
-                    widget.child,
+                    widget.child!,
                   ],
                 );
               }
               return preview;
 
             default:
-              throw new AssertionError("${details.connectionState} not supported.");
+              throw AssertionError("${details.connectionState} not supported.");
           }
         },
       );
@@ -163,26 +167,26 @@ class QrCameraState extends State<QrCamera> with WidgetsBindingObserver {
 class Preview extends StatelessWidget {
   final double width, height;
   final double targetWidth, targetHeight;
-  final int textureId;
-  final int sensorOrientation;
+  final int? textureId;
+  final int? sensorOrientation;
   final BoxFit fit;
 
   Preview({
-    @required PreviewDetails previewDetails,
-    @required this.targetWidth,
-    @required this.targetHeight,
-    @required this.fit,
-  })  : assert(previewDetails != null),
-        textureId = previewDetails.textureId,
-        width = previewDetails.width.toDouble(),
-        height = previewDetails.height.toDouble(),
-        sensorOrientation = previewDetails.sensorOrientation;
+    required PreviewDetails previewDetails,
+    required this.targetWidth,
+    required this.targetHeight,
+    required this.fit,
+  })   : textureId = previewDetails.textureId,
+        width = previewDetails.width!.toDouble(),
+        height = previewDetails.height!.toDouble(),
+        sensorOrientation = previewDetails.sensorOrientation as int?;
 
   @override
   Widget build(BuildContext context) {
-    return new NativeDeviceOrientationReader(
+    return NativeDeviceOrientationReader(
       builder: (context) {
-        var nativeOrientation = NativeDeviceOrientationReader.orientation(context);
+        var nativeOrientation =
+            NativeDeviceOrientationReader.orientation(context);
 
         int nativeRotation = 0;
         switch (nativeOrientation) {
@@ -203,7 +207,8 @@ class Preview extends StatelessWidget {
             break;
         }
 
-        int rotationCompensation = ((nativeRotation - sensorOrientation + 450) % 360) ~/ 90;
+        int rotationCompensation =
+            ((nativeRotation - sensorOrientation! + 450) % 360) ~/ 90;
 
         double frameHeight = width;
         double frameWidth = height;
@@ -216,7 +221,7 @@ class Preview extends StatelessWidget {
               child: SizedBox(
                 width: frameWidth,
                 height: frameHeight,
-                child: Texture(textureId: textureId),
+                child: Texture(textureId: textureId!),
               ),
             ),
           ),
