@@ -1,13 +1,12 @@
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 class PreviewDetails {
-  num width;
-  num height;
-  num sensorOrientation;
-  int textureId;
+  num? width;
+  num? height;
+  num? sensorOrientation;
+  int? textureId;
 
   PreviewDetails(
       this.width, this.height, this.sensorOrientation, this.textureId);
@@ -37,14 +36,14 @@ const _defaultBarcodeFormats = const [
 class QrMobileVision {
   static const MethodChannel _channel =
       const MethodChannel('com.github.rmtmckenzie/qr_mobile_vision');
-  static QrChannelReader channelReader = new QrChannelReader(_channel);
+  static QrChannelReader channelReader = QrChannelReader(_channel);
 
   //Set target size before starting
   static Future<PreviewDetails> start({
-    @required int width,
-    @required int height,
-    @required QRCodeHandler qrCodeHandler,
-    List<BarcodeFormats> formats = _defaultBarcodeFormats,
+    required int width,
+    required int height,
+    required QRCodeHandler qrCodeHandler,
+    List<BarcodeFormats>? formats = _defaultBarcodeFormats,
   }) async {
     final _formats = formats ?? _defaultBarcodeFormats;
     assert(_formats.length > 0);
@@ -64,13 +63,12 @@ class QrMobileVision {
     // invokeMethod returns Map<dynamic,...> in dart 2.0
     assert(details is Map<dynamic, dynamic>);
 
-    int textureId = details["textureId"];
-    num orientation = details["surfaceOrientation"];
-    num surfaceHeight = details["surfaceHeight"];
-    num surfaceWidth = details["surfaceWidth"];
+    int? textureId = details["textureId"];
+    num? orientation = details["surfaceOrientation"];
+    num? surfaceHeight = details["surfaceHeight"];
+    num? surfaceWidth = details["surfaceWidth"];
 
-    return new PreviewDetails(
-        surfaceWidth, surfaceHeight, orientation, textureId);
+    return PreviewDetails(surfaceWidth, surfaceHeight, orientation, textureId);
   }
 
   static Future stop() {
@@ -82,14 +80,15 @@ class QrMobileVision {
     return _channel.invokeMethod('heartbeat').catchError(print);
   }
 
-  static Future<List<List<int>>> getSupportedSizes() {
-    return _channel.invokeMethod('getSupportedSizes').catchError(print);
+  static Future<List<List<int>>?> getSupportedSizes() {
+    return _channel.invokeMethod('getSupportedSizes').catchError(print)
+        as Future<List<List<int>>?>;
   }
 }
 
 enum FrameRotation { none, ninetyCC, oneeighty, twoseventyCC }
 
-typedef void QRCodeHandler(String qr);
+typedef void QRCodeHandler(String? qr);
 
 class QrChannelReader {
   QrChannelReader(this.channel) {
@@ -98,7 +97,7 @@ class QrChannelReader {
         case 'qrRead':
           if (qrCodeHandler != null) {
             assert(call.arguments is String);
-            qrCodeHandler(call.arguments);
+            qrCodeHandler!(call.arguments);
           }
           break;
         default:
@@ -108,10 +107,10 @@ class QrChannelReader {
     });
   }
 
-  void setQrCodeHandler(QRCodeHandler qrch) {
+  void setQrCodeHandler(QRCodeHandler? qrch) {
     this.qrCodeHandler = qrch;
   }
 
   MethodChannel channel;
-  QRCodeHandler qrCodeHandler;
+  QRCodeHandler? qrCodeHandler;
 }
